Update backend to support decimal ratings (half-stars)

+++ src/backend/MovieChecker.Domain/Models/Dtos/EntryRatingDto.cs
@@ -4,5 +4,5 @@ public record EntryRatingDto(
     int Id,
     int UserId,
     string DisplayName,
-    int Rating
+    decimal Rating
 );
