+++ src/backend/MovieChecker.Domain/Models/Dtos/MemberRatingDto.cs
@@ -1,3 +1,3 @@
 namespace MovieChecker.Domain.Models.Dtos;
 
-public record MemberRatingDto(int UserId, string DisplayName, int AverageRating, int TotalRated);
+public record MemberRatingDto(int UserId, string DisplayName, decimal AverageRating, int TotalRated);
