+++ src/backend/MovieChecker.Domain/Models/Dtos/UserRatingInput.cs
@@ -1,3 +1,3 @@
 namespace MovieChecker.Domain.Models.Dtos;
 
-public record UserRatingInput(int UserId, int Rating);
+public record UserRatingInput(int UserId, decimal Rating);
